requirements.txt
tqdm
openai
scikit-learn
pygame
retrying
json5
python-dotenv
requests